Feat : Css 반응형 구현, Fix : 컴포넌트 분리
