+++ src/component/Task/Task.js
@@ -3,6 +3,9 @@ import { v4 as uuidv4 } from 'uuid'; // 고유 ID 생성을 위해 uuid 사용
 import styled from 'styled-components';
 import { getTasksFromLocalStorage, saveTasksToLocalStorage } from '../../utils/localStorage';
 import { CalendarDays, CircleX, CopyPlus, Pencil, Plus, Undo2, X } from 'lucide-react';
+import PendingTask from './components/PendingTask';
+import CompletedTask from './components/CompletedTask';
+import TabIndicator from './components/TabIndicator';
 
 
 function Task({ selectedDate, handleCloseModal }) {
@@ -10,19 +13,6 @@ function Task({ selectedDate, handleCloseModal }) {
     const [newTask, setNewTask] = useState('');
     const [addModal, setAddModal] = useState(false);
     const [activeIndex, setActiveIndex] = useState(0);
-
-
-    // 선택된 날짜에 따른 투두리스트 불러오기
-    useEffect(() => {
-        const formattedDate = selectedDate.toISOString().split('T')[0]; // YYYY-MM-DD 형식
-        const loadedTasks = getTasksFromLocalStorage(formattedDate);
-
-        // Map으로 taskId와 task를 매핑
-        const newTaskMap = new Map();
-        loadedTasks.forEach(task => newTaskMap.set(task.id, task));
-
-        setTasks(newTaskMap);
-    }, [selectedDate]);
 
     // 요일 풀네임 가져오기
     const getDayName = (date) => {
@@ -50,25 +40,6 @@ function Task({ selectedDate, handleCloseModal }) {
         saveTasksToLocalStorage(formattedDate, Array.from(tasks.values()));
         setNewTask('');
     };
-    
-    const handleCompleteTask = (taskId) => {
-        const formattedDate = selectedDate.toISOString().split('T')[0];
-        if (tasks.has(taskId)) {
-            if (window.confirm('이 할 일을 완료하였습니까?')) {  // 완료 확인
-                const task = tasks.get(taskId); // 기존 task 복사
-                const updatedTask = {
-                    ...task,
-                    completed: !task.completed,
-                    completedTime: !task.completed ? new window.Date().toLocaleTimeString() : null,
-                };
-            
-                tasks.set(taskId, updatedTask); // 기존 Map 수정
-            
-                setTasks(new Map(tasks)); // 새로운 Map을 생성하지 않고 기존 Map을 수정한 후 상태를 업데이트
-                saveTasksToLocalStorage(formattedDate, Array.from(tasks.values()));
-            }
-        }
-      };
       const handleDeleteTask = (taskId) => {
         const formattedDate = selectedDate.toISOString().split('T')[0];
         
@@ -93,78 +64,30 @@ function Task({ selectedDate, handleCloseModal }) {
                 </Date>
                 <TaskList>
                     <TabContainer>
-                        {/* Pending Tasks 탭 */}
-                        <RadioInput
-                            type="radio"
-                            id="tab1"
-                            name="tab"
-                            checked={activeIndex === 0}
-                            onChange={() => setActiveIndex(0)}
-                        />
-                        <TabLabel htmlFor="tab1">Pending Tasks</TabLabel>
-                        <ActiveTabIndicator style={{ left: `calc(${activeIndex * 100}% / 2)` }} />
-
-                        {/* Completed Tasks 탭 */}
-                        <RadioInput
-                            type="radio"
-                            id="tab2"
-                            name="tab"
-                            checked={activeIndex === 1}
-                            onChange={() => setActiveIndex(1)}
-                        />
-                        <TabLabel htmlFor="tab2">Completed Tasks</TabLabel>
+                        <TabIndicator 
+                            activeIndex={activeIndex} 
+                            setActiveIndex={setActiveIndex} 
+                            />
                     </TabContainer>
                     <TabContent>
                         {/* Pending Tasks 컨텐츠 */}
                         {activeIndex === 0 && (
-                        <TaskItemsContainer>
-                            {Array.from(tasks.values())
-                            .filter((task) => !task.completed)
-                            .map((task) => (
-                                <TaskItem
-                                    key={task.id}
-                                    onClick={() => handleCompleteTask(task.id)}
-                                    className={task.completed ? 'completed' : ''}
-                                >
-                                    <p className='task-text'>
-                                        {task.text}                                        
-                                    </p>
-                                    <p className='task-time'>
-                                        <span className='util-btn'>
-                                            <button>Edit</button>
-                                            <button onClick={()=>handleDeleteTask(task.id)}>Delete</button>
-                                        </span>
-                                        <span>작성시간 : {task.createdTime}</span>
-                                    </p>
-                                </TaskItem>
-                            ))}
-                        </TaskItemsContainer>
+                            <PendingTask 
+                                tasks={tasks} 
+                                setTasks={setTasks} 
+                                handleDeleteTask={handleDeleteTask}
+                                selectedDate={selectedDate}
+                                />
                         )}
 
                         {/* Completed Tasks 컨텐츠 */}
                         {activeIndex === 1 && (
-                        <TaskItemsContainer>
-                            {Array.from(tasks.values())
-                            .filter((task) => task.completed)
-                            .map((task) => (
-                                <TaskItem
-                                    key={task.id}
-                                    className={task.completed ? 'completed' : ''}
-                                    onClick={() => handleCompleteTask(task.id)}
-                                >
-                                    <p className='task-text'>
-                                        {task.text}                                        
-                                    </p>
-                                    <p className='task-time'>
-                                        <span className='util-btn'>
-                                            <button>Edit</button>
-                                            <button onClick={()=>handleDeleteTask(task.id)}>Delete</button>
-                                        </span>
-                                        <span>완료시간 : {task.createdTime}</span>
-                                    </p>
-                                </TaskItem>
-                            ))}
-                        </TaskItemsContainer>
+                            <CompletedTask 
+                                tasks={tasks} 
+                                setTasks={setTasks} 
+                                handleDeleteTask={handleDeleteTask}
+                                selectedDate={selectedDate}
+                                />
                         )}
                     </TabContent>
                 </TaskList>
@@ -176,6 +99,14 @@ function Task({ selectedDate, handleCloseModal }) {
                                 value={newTask}
                                 onChange={(e) => setNewTask(e.target.value)}
                                 placeholder="Add a new task"
+                                required
+                                autoComplete='on'
+                                onKeyDown={(e) => {
+                                    if (e.key === 'Enter') {
+                                        e.preventDefault();  // 기본 Enter 키 동작 막기
+                                        handleAddTask();
+                                    }
+                                }}
                             />
                             <span onClick={handleAddTask}>
                                 <CopyPlus/>
@@ -195,8 +126,6 @@ function Task({ selectedDate, handleCloseModal }) {
 
                         </BtnSpan>
                     }
-                    
-                    
                 </TaskFooter>
             </Wrapper>
         </>
@@ -226,81 +155,24 @@ const TaskList = styled.div`
     width: 100%;
     height: 70%;
 `;
-
-const TaskColumn = styled.div`
-    width: 50%;
-    height: 100%;
-    padding-bottom: 10px;
-    overflow: hidden;
+const TabContainer = styled.div`
+  display: flex;
+  background-color: whitesmoke;
+  border-radius: 25px;
+  overflow: hidden;
+  position: relative;
+  margin-top: 10px;
+  width: 80%;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 `;
-const TaskHeader = styled.h3`
-    margin: 0;
-    padding: 10px;
-    border-bottom: 1px solid #ddd; /* 구분선 추가 */
-    font-size: 1.2rem;
-    text-align: center;
-`;
-
-const TaskItemsContainer = styled.div`
-    margin-top: 10px;
-    padding: 10px;
-    display: flex;
-    flex-direction: column;
-    height: calc(100%); /* 컨테이너 높이에서 여유 공간을 확보 */
-    overflow-y: auto; /* 항목들에 대한 스크롤 */
-    /* 스크롤바 숨기기 */
-    &::-webkit-scrollbar {
-        display: none; /* 크롬, 사파리, 엣지에서 스크롤바 숨기기 */
-    }
-    -ms-overflow-style: none;  /* IE와 Edge에서 스크롤바 숨기기 */
-    scrollbar-width: none;  /* Firefox에서 스크롤바 숨기기 */
-`;
-
-
-const TaskItem = styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    width: auto;
-    height: auto;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    padding: 10px;
-    margin-bottom: 10px; 
-    &:hover {
-        background-color: #f0f0f0;
-    }
-    transition: all 0.5s ease; /* 트랜지션 추가 */
-    cursor: pointer;
-    .task-text {
-        margin: 0;
-        font-size: 1rem;
-        display: flex;
-        justify-content: space-between;
-        overflow-wrap: break-word;
-    }
-    .task-time {
-        display: flex;
-        justify-content: flex-end;
-        gap: 5px;
-        margin: 0;
-        font-size: 0.8rem;
-        color: gray;
-    }
-    .util-btn {
-        display: flex;
-        cursor: pointer;
-        button {
-            border: none;
-            background-color: transparent;
-            color: #007bff;
-            cursor: pointer;
-            &:hover {
-                text-decoration: underline;
-            }
-        }
-    }
-
+const TabContent = styled.div`
+  width: 80%;
+  height: 85%;
+  margin: auto;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  position: relative;
 `;
 const TaskFooter = styled.div`
     display: flex;
@@ -364,68 +236,6 @@ const BtnSpan = styled.div`
         }
     }
 `;  
-const TabContainer = styled.div`
-  display: flex;
-  background-color: whitesmoke;
-  border-radius: 25px;
-  overflow: hidden;
-  position: relative;
-  margin-top: 10px;
-  width: 80%;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-`;
 
-const RadioInput = styled.input`
-  display: none;
-
-  &:checked + label {
-    color: white;
-  }
-
-  &:checked + label + div {
-    left: ${({ index }) => `calc(${index * 100}% / 2)`};
-  }
-`;
-
-const TabLabel = styled.label`
-  flex: 1;
-  padding: 15px 20px;
-  background-color: transparent;
-  cursor: pointer;
-  font-size: 16px;
-  color: #333;
-  text-align: center;
-  position: relative;
-  z-index: 1;
-  transition: color 0.5s ease;
-
-  &:hover {
-    background-color: rgba(1, 0, 0, 0.01);
-  }
-  @media (min-width: 360px) {
-    font-size: 12px;
-  } 
-`;
-
-const ActiveTabIndicator = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: calc(100% / 2);
-  height: 100%;
-  background-color: rgb(62, 76, 247);
-  border-radius: 25px;
-  transition: left 0.3s ease;
-  z-index: 0;
-`;
-const TabContent = styled.div`
-  width: 80%;
-  height: 85%;
-  margin: auto;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  position: relative;
-`;
 
 export default Task;
